# web/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "tsundoku API",
    "description": "Manga discovery service. Read endpoints (`GET`) are public unless `auth.read_requires_auth = true`. Write endpoints (`POST`) always require the admin bearer.",
    "contact": {
      "name": "Sylvain Cau",
      "email": "sylvain.cau@gmail.com"
    },
    "license": {
      "name": "MIT OR Apache-2.0",
      "identifier": "MIT OR Apache-2.0"
    },
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/providers/{id}/refresh-cache": {
      "post": {
        "tags": [
          "providers"
        ],
        "summary": "Trigger an offline-cache refresh for the named provider. Same locking\nsemantics as the scheduled refresh job; an in-flight refresh causes\nthis request to report `skipped = true` without spawning a duplicate.",
        "operationId": "refresh_cache",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Provider id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "202": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RefreshResponse"
                }
              }
            }
          },
          "404": {
            "description": "No provider with that id registered"
          }
        },
        "security": [
          {
            "admin": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "RefreshResponse": {
        "type": "object",
        "required": [
          "provider",
          "triggered",
          "skipped"
        ],
        "properties": {
          "provider": {
            "type": "string"
          },
          "skipped": {
            "type": "boolean",
            "description": "`false` when a refresh is already in flight; the request is a no-op."
          },
          "triggered": {
            "type": "boolean"
          }
        }
      },
...
    },
    "securitySchemes": {
      "admin": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "opaque",
        "description": "Admin bearer token (auth.admin_token in config)."
      }
    }
  },
...
}